requirements.txt
python-dotenv==1.0.0
fastapi==0.108.0
fastapi-restful==0.5.0
langchain==0.1.0
langchain-openai==0.0.2.post1
pydantic==2.5.3
pinecone-client==2.2.4
tiktoken==0.5.2
openai==1.8.0
uvicorn[standard]==0.20.0
gunicorn==20.1.0